backend: bugfix for turing machine related endpoints

+++ backend/TuringMachine.Backend/TuringMachine.Backend.Server/Database/Entity/UiLabels/MachineLabels/MachineBoxLabel.cs
@@ -2,13 +2,13 @@
 {
     internal class MachineBoxLabel
     {
-        public Guid  MachineLabelID { get; set; }
-        public float StartX         { get; set; }
-        public float StartY         { get; set; }
-        public float Width          { get; set; }
-        public float Height         { get; set; }
-        public int   Color          { get; set; }
-        public byte  LabelIndex     { get; set; }
+        public Guid   MachineLabelID { get; set; }
+        public double StartX         { get; set; }
+        public double StartY         { get; set; }
+        public double Width          { get; set; }
+        public double Height         { get; set; }
+        public int    Color          { get; set; }
+        public byte   LabelIndex     { get; set; }
 
 
         #region Relationship
