+++ backend/TuringMachine.Backend/TuringMachine.Backend.Server/Database/Entity/UiLabels/MachineLabels/TextLabel.cs
@@ -3,8 +3,8 @@
     internal class TextLabel
     {
         public Guid   MachineLabelID { get; set; }
-        public float  PosX           { get; set; }
-        public float  PosY           { get; set; }
+        public double PosX           { get; set; }
+        public double PosY           { get; set; }
         public string Value          { get; set; }
         public byte   LabelIndex     { get; set; }
